+++ common/src/states/leap_melee.rs
@@ -1,114 +1,197 @@
 use crate::{
     comp::{Attacking, CharacterState, StateUpdate},
-    states::utils::*,
+    states::utils::{StageSection, *},
     sys::character_behavior::{CharacterBehavior, JoinData},
 };
 use serde::{Deserialize, Serialize};
 use std::time::Duration;
 use vek::Vec3;
 
-const LEAP_SPEED: f32 = 24.0;
-
-#[derive(Copy, Clone, Debug, PartialEq, Serialize, Deserialize, Eq, Hash)]
-pub struct Data {
+/// Separated out to condense update portions of character state
+#[derive(Copy, Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub struct StaticData {
     /// How long the state is moving
     pub movement_duration: Duration,
     /// How long until state should deal damage
     pub buildup_duration: Duration,
+    /// How long the weapon swings
+    pub swing_duration: Duration,
     /// How long the state has until exiting
     pub recover_duration: Duration,
     /// Base damage
     pub base_damage: u32,
+    /// Knockback
+    pub knockback: f32,
+    /// Max range
+    pub range: f32,
+    /// Max angle (45.0 will give you a 90.0 angle window)
+    pub max_angle: f32,
+    /// Affects how far forward the player leaps
+    pub forward_leap_strength: f32,
+    /// Affects how high the player leaps
+    pub vertical_leap_strength: f32,
+}
+
+#[derive(Copy, Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub struct Data {
+    /// Struct containing data that does not change over the course of the
+    /// character state
+    pub static_data: StaticData,
+    /// Timer for each stage
+    pub timer: Duration,
+    /// What section the character stage is in
+    pub stage_section: StageSection,
     /// Whether the attack can deal more damage
     pub exhausted: bool,
-    pub initialize: bool,
 }
 
 impl CharacterBehavior for Data {
     fn behavior(&self, data: &JoinData) -> StateUpdate {
         let mut update = StateUpdate::from(data);
 
-        if self.initialize {
-            update.vel.0 = *data.inputs.look_dir * 20.0;
-            if let Some(dir) = Vec3::from(data.inputs.look_dir.xy()).try_normalized() {
-                update.ori.0 = dir.into();
-            }
-        }
+        handle_move(data, &mut update, 0.3);
+        handle_jump(data, &mut update);
 
-        if self.movement_duration != Duration::default() {
-            // Jumping
-            update.vel.0 = Vec3::new(data.inputs.look_dir.x, data.inputs.look_dir.y, 8.0)
-                * ((self.movement_duration.as_millis() as f32) / 250.0)
-                + (update.vel.0 * Vec3::new(2.0, 2.0, 0.0)
-                    + 0.25 * data.inputs.move_dir.try_normalized().unwrap_or_default())
-                .try_normalized()
-                .unwrap_or_default()
-                    * LEAP_SPEED
-                    * (1.0 - data.inputs.look_dir.z.abs());
+        match self.stage_section {
+            StageSection::Buildup => {
+                if self.timer < self.static_data.buildup_duration {
+                    // Buildup
+                    update.character = CharacterState::LeapMelee(Data {
+                        static_data: self.static_data,
+                        timer: self
+                            .timer
+                            .checked_add(Duration::from_secs_f32(data.dt.0))
+                            .unwrap_or_default(),
+                        stage_section: self.stage_section,
+                        exhausted: self.exhausted,
+                    });
+                } else {
+                    // Transitions to leap portion of state
+                    update.character = CharacterState::LeapMelee(Data {
+                        static_data: self.static_data,
+                        timer: Duration::default(),
+                        stage_section: StageSection::Movement,
+                        exhausted: self.exhausted,
+                    });
+                }
+            },
+            StageSection::Movement => {
+                // Jumping
+                update.vel.0 = Vec3::new(
+                    data.inputs.look_dir.x,
+                    data.inputs.look_dir.y,
+                    self.static_data.vertical_leap_strength,
+                ) * 2.0
+                    * (1.0
+                        - self.timer.as_secs_f32()
+                            / self.static_data.movement_duration.as_secs_f32())
+                    + (update.vel.0 * Vec3::new(2.0, 2.0, 0.0)
+                        + 0.25 * data.inputs.move_dir.try_normalized().unwrap_or_default())
+                    .try_normalized()
+                    .unwrap_or_default()
+                        * self.static_data.forward_leap_strength
+                        * (1.0 - data.inputs.look_dir.z.abs());
 
-            update.character = CharacterState::LeapMelee(Data {
-                movement_duration: self
-                    .movement_duration
-                    .checked_sub(Duration::from_secs_f32(data.dt.0))
-                    .unwrap_or_default(),
-                buildup_duration: self.buildup_duration,
-                recover_duration: self.recover_duration,
-                base_damage: self.base_damage,
-                exhausted: false,
-                initialize: false,
-            });
-        } else if self.buildup_duration != Duration::default() && !data.physics.on_ground {
-            // Falling
-            update.character = CharacterState::LeapMelee(Data {
-                movement_duration: Duration::default(),
-                buildup_duration: self
-                    .buildup_duration
-                    .checked_sub(Duration::from_secs_f32(data.dt.0))
-                    .unwrap_or_default(),
-                recover_duration: self.recover_duration,
-                base_damage: self.base_damage,
-                exhausted: false,
-                initialize: false,
-            });
-        } else if !self.exhausted {
-            // Hit attempt
-            data.updater.insert(data.entity, Attacking {
-                base_damage: self.base_damage,
-                base_heal: 0,
-                range: 4.5,
-                max_angle: 360_f32.to_radians(),
-                applied: false,
-                hit_count: 0,
-                knockback: 25.0,
-            });
+                if self.timer < self.static_data.movement_duration {
+                    // Movement duration
+                    update.character = CharacterState::LeapMelee(Data {
+                        static_data: self.static_data,
+                        timer: self
+                            .timer
+                            .checked_add(Duration::from_secs_f32(data.dt.0))
+                            .unwrap_or_default(),
+                        stage_section: self.stage_section,
+                        exhausted: self.exhausted,
+                    });
+                } else {
+                    // Transitions to swing portion of state
+                    update.character = CharacterState::LeapMelee(Data {
+                        static_data: self.static_data,
+                        timer: Duration::default(),
+                        stage_section: StageSection::Swing,
+                        exhausted: self.exhausted,
+                    });
+                }
+            },
+            StageSection::Swing => {
+                if self.timer < self.static_data.swing_duration {
+                    // Swings weapons
+                    update.character = CharacterState::LeapMelee(Data {
+                        static_data: self.static_data,
+                        timer: self
+                            .timer
+                            .checked_add(Duration::from_secs_f32(data.dt.0))
+                            .unwrap_or_default(),
+                        stage_section: self.stage_section,
+                        exhausted: self.exhausted,
+                    });
+                } else {
+                    // Transitions to recover portion
+                    update.character = CharacterState::LeapMelee(Data {
+                        static_data: self.static_data,
+                        timer: Duration::default(),
+                        stage_section: StageSection::Recover,
+                        exhausted: self.exhausted,
+                    });
+                }
+            },
+            StageSection::Recover => {
+                if !data.physics.on_ground {
+                    // Falls
+                    update.character = CharacterState::LeapMelee(Data {
+                        static_data: self.static_data,
+                        timer: self
+                            .timer
+                            .checked_add(Duration::from_secs_f32(data.dt.0))
+                            .unwrap_or_default(),
+                        stage_section: self.stage_section,
+                        exhausted: self.exhausted,
+                    });
+                } else if !self.exhausted {
+                    // Hit attempt
+                    data.updater.insert(data.entity, Attacking {
+                        base_damage: self.static_data.base_damage,
+                        base_heal: 0,
+                        range: self.static_data.range,
+                        max_angle: self.static_data.max_angle.to_radians(),
+                        applied: false,
+                        hit_count: 0,
+                        knockback: self.static_data.knockback,
+                    });
 
-            update.character = CharacterState::LeapMelee(Data {
-                movement_duration: self.movement_duration,
-                buildup_duration: Duration::default(),
-                recover_duration: self.recover_duration,
-                base_damage: self.base_damage,
-                exhausted: true,
-                initialize: false,
-            });
-        } else if self.recover_duration != Duration::default() {
-            // Recovery
-            handle_move(data, &mut update, 0.7);
-            update.character = CharacterState::LeapMelee(Data {
-                movement_duration: self.movement_duration,
-                buildup_duration: self.buildup_duration,
-                recover_duration: self
-                    .recover_duration
-                    .checked_sub(Duration::from_secs_f32(data.dt.0))
-                    .unwrap_or_default(),
-                base_damage: self.base_damage,
-                exhausted: true,
-                initialize: false,
-            });
-        } else {
-            // Done
-            update.character = CharacterState::Wielding;
-            // Make sure attack component is removed
-            data.updater.remove::<Attacking>(data.entity);
+                    update.character = CharacterState::LeapMelee(Data {
+                        static_data: self.static_data,
+                        timer: self
+                            .timer
+                            .checked_add(Duration::from_secs_f32(data.dt.0))
+                            .unwrap_or_default(),
+                        stage_section: self.stage_section,
+                        exhausted: true,
+                    });
+                } else if self.timer < self.static_data.recover_duration {
+                    // Recovers
+                    update.character = CharacterState::LeapMelee(Data {
+                        static_data: self.static_data,
+                        timer: self
+                            .timer
+                            .checked_add(Duration::from_secs_f32(data.dt.0))
+                            .unwrap_or_default(),
+                        stage_section: self.stage_section,
+                        exhausted: self.exhausted,
+                    });
+                } else {
+                    // Done
+                    update.character = CharacterState::Wielding;
+                    // Make sure attack component is removed
+                    data.updater.remove::<Attacking>(data.entity);
+                }
+            },
+            _ => {
+                // If it somehow ends up in an incorrect stage section
+                update.character = CharacterState::Wielding;
+                // Make sure attack component is removed
+                data.updater.remove::<Attacking>(data.entity);
+            },
         }
 
         update